hwclock: use RTC_PARAM_GET and RTC_VL_READ for ifdefs
Replace broad #ifndef __GNU__ guards with feature-specific #ifdef
RTC_PARAM_GET and #ifdef RTC_VL_READ to guard the code that depends
on these ioctls.

Move the fallback definitions for RTC_PARAM_GET, RTC_PARAM_SET,
RTC_VL_READ, and RTC_VL_CLR from hwclock-rtc.c to hwclock.h under
feature macros consistently.
